fix: meta usage, rm usage, file DNE
This commit brings in fixes for 3 things.
(i) createMetaData function's order of parameters is the reverse of
order in use. We need to stick to that.
(ii) removeFile(destination) does not make sense since the passed
parameter is a directory.
(iii) the case for a non-existent file in .meta was not handled.