feat: encrypt and decrypt file, hash pwds
This commit brings in all cryptographic features of this program. We
use AES256 in CTR mode to encrypt and decrypt files. We take user's
password and apply SHA256 over it to get a 256-bit key, used by AES.
We again hash this key and get the first 128-bits of the resultant hash
as our IV for AES. This puts the responsbility on the user to provide a
strong password. With the possibility of a password manager coming up
in rwxLife software set, this shouldn't be a problem.